+++ src/main/java/cz/aegis/corebanking/exception/ClientNotFoundException.java
@@ -1,0 +1,8 @@
+package cz.aegis.corebanking.exception;
+
+public class ClientNotFoundException extends RuntimeException {
+
+    public ClientNotFoundException(Long id) {
+        super("Client with ID " + id + " not found.");
+    }
+}